Create 30. Basic exception handling
Objective:
Create a program, which asks the user for input and tries to convert it to an integer value. 
Rules:
If the conversion happens without problems, the program prints "The input was suitable!". If the user gives something which does not convert, like letters or special characters, the program avoids the error with an exception handler and prints "The input was malformed."