Create 31. Catching various types of errors
Objective:
Create a program which prompts the user for a file name. Based on user input, open the given file and read the contents into one big string. Then convert this string to an integer and divides the number 1000 with the number. Finally, print out the result from the division.
Rules:
No matter what the user input is, the program works. If the file cannot be found, the program prints "There seems to be no file with that name.", if the conversion fails, "The file contents are unsuitable.", in other errors "There was a problem with the program." or if everything went correctly, "The result was [result].". In any case (besides KeyboardInterruption with Ctrl-C), the program should be impossible to break with user input.