Create 35. Basic shopping list
Objective:
Create a small grocery shopping list that allows the user to add, remove, print the list and quit.
Rules:
If the user adds something to the list, the program asks "What will be added?" and saves it as the last item in the list. If the user decides to remove something, the program informs the user about how many items there are on the list (There are [number] items in the list.") and prompts the user for the removed item ("Which item is deleted?: "). If the user selects 0, the first item is removed. When the user quits, the final list is printed for the user "The following items remain in the list:" followed by the remaining items one per line. If the user selects anything outside the options, including when deleting items, the program responds "Incorrect selection."