Create 37. Notebook: using pickle
Objective:
The idea is to build a notebook which applies the Python data structure list as a note manipulation method when the program is executed and saves the data in a bitwise mode with pickle. 
Rules:
The program has the following features:
A) It is possible to add notes, with similar timestamps as previous ones.
B) It is possible to edit a note by selecting it from the list, and rewriting it with new data.
C) It is possible to delete notes separately based on the selection and
D) Notebook loads an existing notebook file from the file "notebook.dat" if such a file exists.