Create 39. Create a method
Objective;
Create a class with two attributes, teamcolor and points. Now, add two methods called "tellscore" and "goal". 
Rules:
When the method "tellscore" is called, the objects prints out "I am [teamcolor], we have [points] points!", and if the method goal is called, the object adds one point to the attribute points. Also, make the attribute points a private attribute, and to accommodate that change, make the necessary changes to get the class to work properly.