Create 40. Initializing a class using  _init_
Objective:
Adjust the class by adding initialization data and a docstring.
Rules:
Add a docstring "Player-class: stores data on team colors and points.". After this, add an initializing method __init__ to the class, and make it prompt the user for a new player color with the message"What color do I get?: ".